requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
websockets>=12.0
pydantic>=2.5.0
numpy>=1.24.0
pydub>=0.25.1
python-dotenv>=1.0.0
aiofiles>=23.0.0
